requirements.txt
asgiref==3.11.1
Django==5.2.11
django-filter==25.2
django-htmx==1.27.0
djangorestframework==3.16.1
Markdown==3.10.2
pillow==12.1.1
sqlparse==0.5.5
typing_extensions==4.15.0
